% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_apex.R
\name{cv.apex}
\alias{cv.apex}
\title{Détection des apex}
\usage{
cv.apex(h, lim_h_rege, pente = 0.07, intercept = 2)
}
\arguments{
\item{h}{MNH spatraster}

\item{lim_h_rege}{hauteur minimale des houppiers}

\item{pente}{pente de la fct de surface minimale}

\item{intercept}{intercept}
}
\value{
spatial points
}
\description{
Détection des apex
}
