% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_carto_evo.R
\name{cv.carto_evo}
\alias{cv.carto_evo}
\title{Cartographie des différences entre 2 mnh de dates différntes}
\usage{
cv.carto_evo(
  shp_ug,
  mnh0,
  mnh1,
  an0,
  an1,
  lim_h_rege = 3,
  lim_h_perche = 15,
  lim_dh_rege = 0.2,
  lim_dh_perche = 0.5,
  lim_dh_futaie = 0.3,
  max_acc = 1,
  max_area_perche = 30,
  seuil_dtpi = 2,
  desserte = NULL,
  champ_id,
  dos_dest
)
}
\arguments{
\item{shp_ug}{polygone sf}

\item{mnh0}{mnh an 0}

\item{mnh1}{mnh an 1}

\item{an0}{année premier mnh}

\item{an1}{année second mnh}

\item{lim_h_rege}{seuil hauteur catégorie régénération}

\item{lim_h_perche}{id pour les jeunes peuplemments}

\item{lim_dh_rege}{seuil acc mini régé}

\item{lim_dh_perche}{seuil acc annuel fort/faible pour les ppts jeunes}

\item{lim_dh_futaie}{id pour les ppts adultes}

\item{max_acc}{accroissement max en m/an au-delà duquel une fermeture latérale du couvert est vraisemblable}

\item{max_area_perche}{surf max d'un houppier de h < lim_h_perche pour qu'il soit considéré comme encore jeune}

\item{seuil_dtpi}{différence de TPI au-delà duquel le pixel est considéré comme en fermeture latérale du couvert}

\item{desserte}{sf des dessertes avec champ 'largeur'}

\item{champ_id}{champ d'identifiant unique de shp,}

\item{dos_dest}{dossier d'écriture des tifs}
}
\value{
liste: raster et table
}
\description{
Cartographie des différences entre 2 mnh de dates différntes
}
