% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_color_evo.R
\name{cv.color_evo}
\alias{cv.color_evo}
\title{catégorise et colore les rasters d'évolution}
\usage{
cv.color_evo(r, plot = TRUE, shp = NULL)
}
\arguments{
\item{r}{spatraster d'évolution du couvert}

\item{plot}{TRUE pour éditer une carte}

\item{shp}{(fac) polygones à ajouter à la carte}
}
\value{
spatraster
}
\description{
catégorise et colore les rasters d'évolution
}
