% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_fct_volume.R
\name{cv.fct_volume}
\alias{cv.fct_volume}
\title{Fonction V = f(mnh)}
\usage{
cv.fct_volume(
  an,
  path_dendro = data_conf("path_dendro"),
  path_mnh = data_conf("path_mnh_ts"),
  fh = 10
)
}
\arguments{
\item{an}{année MNH utilisée}

\item{path_dendro}{chemin rasters dendrométriques}

\item{path_mnh}{chemin raster mnh (série temporelle)}

\item{fh}{FH}
}
\value{
fonction
}
\description{
Fonction V = f(mnh)
}
