% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_carto_evo_frt.R
\name{cv_carto_evo_frt}
\alias{cv_carto_evo_frt}
\title{Cartographie des différences entre 2 mnh de dates différntes sur une forêt entière}
\usage{
cv_carto_evo_frt(
  shp = oiseauData::data_conf("shp"),
  mnh_ts = terra::rast(oiseauData::data_conf("path_mnh_ts")),
  an0 = oiseauData::data_conf("an0"),
  an1 = oiseauData::data_conf("an1"),
  lim_h_rege = oiseauData::data_conf("lim_h_rege"),
  lim_h_perche = oiseauData::data_conf("lim_h_perche"),
  lim_dh_rege = oiseauData::data_conf("lim_dh_rege"),
  lim_dh_perche = oiseauData::data_conf("lim_dh_perche"),
  lim_dh_futaie = oiseauData::data_conf("lim_dh_futaie"),
  max_acc = oiseauData::data_conf("max_acc"),
  max_area_perche = oiseauData::data_conf("max_area_perche"),
  seuil_dtpi = oiseauData::data_conf("seuil_dtpi"),
  path_desserte = oiseauData::data_conf("dos_shp_desserte"),
  champ_id = "id",
  w = oiseauData::data_conf("w"),
  buffer = oiseauData::data_conf("buffer"),
  dos_evo = oiseauData::data_conf("dos_evo"),
  path_mnt = oiseauData::data_conf("path_mnt"),
  force_recalc = FALSE
)
}
\arguments{
\item{shp}{sf de la forêt, avec champ 'CCOD_PRF}

\item{an0}{année premier mnh}

\item{an1}{année second mnh}

\item{lim_h_rege}{seuil hauteur catégorie régénération}

\item{lim_h_perche}{id pour les jeunes peuplemments}

\item{lim_dh_rege}{seuil acc mini régé}

\item{lim_dh_perche}{seuil acc annuel fort/faible pour les ppts jeunes}

\item{lim_dh_futaie}{id pour les ppts adultes}

\item{max_acc}{accroissement max en m/an au-delà duquel une fermeture latérale du couvert est vraisemblable}

\item{max_area_perche}{surf max d'un houppier de h < lim_h_perche pour qu'il soit considéré comme encore jeune}

\item{seuil_dtpi}{différence de TPI au-delà duquel le pixel est considéré comme en fermeture latérale du couvert}

\item{path_desserte}{chemin du shapefile des dessertes avec champ 'largeur'}

\item{champ_id}{champ d'identifiant unique de shp,}

\item{w}{largueur de la fenêtre d'aggrégation (impaire)}

\item{buffer}{tampon de traitement en m autour de shp,}

\item{force_recalc}{TRUE pour calculer l'évolution même si la donnée existe déjà}

\item{mnh0}{mnh an 0}

\item{mnh1}{mnh an 1}

\item{dir}{chemin du dossier du projet. si non NULL, utilise oiseauData::data_conf pour les arguments précédents}
}
\value{
liste: spatRaster et table
}
\description{
Cartographie des différences entre 2 mnh de dates différntes sur une forêt entière
}
