% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mask_ligne_edf.R
\name{cv_mask_ligne_edf}
\alias{cv_mask_ligne_edf}
\title{Masque des emprises EDF}
\usage{
cv_mask_ligne_edf(dest)
}
\arguments{
\item{dest}{fichier d'écriture du raster}
}
\value{
spatraster 1 = emprise, NA ailleurs
}
\description{
Masque des emprises EDF
}
