% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_surfaces_evo.R
\name{cv_surfaces_evo}
\alias{cv_surfaces_evo}
\title{Bilan des évolutions de surfaces de couverts}
\usage{
cv_surfaces_evo(conf = readRDS(file.path(.dir, "conf.rds")), force = TRUE)
}
\arguments{
\item{conf}{conf}

\item{evo}{liste evo}
}
\value{
liste: table et leaflet
}
\description{
Bilan des évolutions de surfaces de couverts
}
