% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_volume_ans.R
\name{cv_volume_ans}
\alias{cv_volume_ans}
\title{Table des volumes disparus après la date du dernier MNH}
\usage{
cv_volume_ans(
  shp = oiseauData::data_conf("shp"),
  path_mnh = oiseauData::data_conf("path_mnh_ts"),
  path_mnh_mask_ts = oiseauData::data_conf("path_mnh_mask_ts"),
  path_emprise_edf.fac = oiseauData::data_conf("path_emprise_edf")
)
}
\value{
table
}
\description{
Table des volumes disparus après la date du dernier MNH
}
