% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_volume_ini.R
\name{cv_volume_ini}
\alias{cv_volume_ini}
\title{Table des volumes à la date du dernier MNH}
\usage{
cv_volume_ini(
  shp = oiseauData::data_conf("shp"),
  an0 = oiseauData::data_conf("an0"),
  an1 = oiseauData::data_conf("an1"),
  path_mnh = oiseauData::data_conf("path_mnh_ts"),
  path_emprise_edf.fac = oiseauData::data_conf("path_emprise_edf"),
  leaflet.fac = FALSE
)
}
\arguments{
\item{leaflet}{TRUE pour afficher la carte des résultats}
}
\value{
table écrite dans conf$tab_volumes_ini
}
\description{
volumes aux dates des MNH disponible, prélèvements et accroissement
}
