% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_volume_post.R
\name{data_volume_post}
\alias{data_volume_post}
\title{Table des volumes disparus après la date du dernier MNH}
\usage{
data_volume_post(
  shp = data_conf("shp"),
  path_mnh = data_conf("path_mnh_ts"),
  path_deads_ts = data_conf("path_deads_ts"),
  path_emprise_edf.fac = data_conf("path_emprise_edf")
)
}
\value{
table
}
\description{
Table des volumes disparus après la date du dernier MNH
}
