% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_insolation.R
\name{cv_insolation}
\alias{cv_insolation}
\title{Insolation reçue à la hauteur du seuil de la strate basse}
\usage{
cv_insolation(
  dos_proj = data_conf("dos_projet"),
  path_mnt = data_conf("path_mnt"),
  path_mnh = data_conf("path_mnh_ts"),
  shp = data_conf("shp"),
  lim_h_rege = data_conf("lim_h_rege"),
  dest_insol = data_conf("path_insolation_ts"),
  resolution = 10,
  force = FALSE
)
}
\arguments{
\item{conf}{conf}

\item{mnh}{liste de mnh}
}
\value{
liste: rasters des 2 années + leaflet + table id + ggplot
}
\description{
Insolation reçue à la hauteur du seuil de la strate basse
}
