% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_trouees.R
\name{cv_trouees}
\alias{cv_trouees}
\title{Identification des trouées à suivre}
\usage{
cv_trouees(
  min_size = oiseauData::data_conf("min_size_trouees"),
  txOuv = oiseauData::data_conf("tx_ouv_trouees"),
  path_evo_renouv = oiseauData::data_conf("path_evo_renouv"),
  path_mnh_mask_ts = oiseauData::data_conf("path_mnh_mask_ts"),
  dest_trouees_suivies = oiseauData::data_conf("dest_trouees_suivies"),
  an0 = oiseauData::data_conf("an0")
)
}
\arguments{
\item{min_size}{surface mini de la trouée en ares}

\item{txOuv}{taux d'ouverture au-dessus duquel le peuplement est concidéré comme trouée}
}
\value{
raster écrit sous "dest_trouees_suivies"
}
\description{
Identification des trouées à suivre
}
